Added js files to modularize the project
