+++ src/index.html
@@ -35,6 +35,5 @@
         <button class="clear-all-btn"> Clear Complete</button>
         <div class="notification"></div>
     </main>
-    <script type="module" src="./index.js"></script>
 </body> 
 </html>
